pyproject.toml
[tool.poetry]
authors = ["Andras Nagy <andrasnagy097@gmail.com>"]
description = "Web application"
name = "server"
version = "0.1.0"

[tool.poetry.dependencies]
SQLAlchemy = "^1.4.46"
SQLAlchemy-Utils = "^0.39.0"
alembic = "^1.9.2"
fastapi = "^0.89.1"
pydantic = "^1.10.4"
python = "^3.10"
python-dotenv = "^0.21.1"
uvicorn = "^0.20.0"

[tool.poetry.dev-dependencies]
black = "^22.12.0"
flake8 = "^6.0.0"
isort = "^5.11.4"
pdoc = "^12.3.1"
pre-commit = "^2.21.0"
pytest = "^7.2"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.black]
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
include = '\.pyi?$'
line-length = 79

[tool.isort]
force_grid_wrap = 0
include_trailing_comma = true
line_length = 79
multi_line_output = 3
profile = "black"
